Remove store file
